Refactor FIAT-based physical mapping into a separate file to facilitate reuse
